+++ src/main/java/com/eatfast/storemeal/repository/StoreMealStatusRepository.java
@@ -25,14 +25,14 @@ public interface StoreMealStatusRepository extends JpaRepository<StoreMealStatus
      * @param mealId  餐點的 ID
      * @return Optional 包裝的 StoreMealStatusEntity 物件，用於優雅地處理可能找不到結果的情況
      */
-    Optional<StoreMealStatusEntity> findByStore_StoreIdAndMeal_MealId(Integer storeId, Long mealId);
+    Optional<StoreMealStatusEntity> findByStore_StoreIdAndMeal_MealId(Long storeId, Long mealId);
 
     /**
      * 查找某個特定門市所有的餐點供應狀態。
      * @param storeId 門市的 ID
      * @return 該門市所有餐點狀態的列表
      */
-    List<StoreMealStatusEntity> findByStore_StoreId(Integer storeId);
+    List<StoreMealStatusEntity> findByStore_StoreId(Long storeId);
     
     /**
      * 查找某個特定餐點在所有門市的供應狀態。
